test(libeval): split trace-collector by behaviour family
Split trace-collector.test.js (518 LOC) into
trace-collector-addline.test.js (the addLine describe) and
trace-collector-output.test.js (toJSON, toText, createTraceCollector).
loadFixture/collectFixture stay with the output family that uses them.
Assertions unchanged; 25 tests preserved.

Spec 0640 § C (large files), design Decision 5.
